+++ basic/update_soft.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
-INSTALL_PATH=$userHome/workspace
+INSTALL_PATH=/home/testuser/workspace
 unzipPath="/home/testuser/workspace/tmp"
-softName=`nameToPath.sh $1`
+[ -n "$1" ] || { echo "update_soft parameter error:no parameter";exit -1; }
+softName=`bash nameToPath.sh $1` || exit $?
 [ -d $INSTALL_PATH/$softName/lib ] || { echo "not found file";exit -1 }
 		cd $INSTALL_PATH/$softName/
 		rm -fr lib releases erts-*
